+++ docs/src/changelog.md
@@ -9,13 +9,18 @@ nav_order: 13
 
 All notable changes to this project will be documented in this section.
 
+## [1.0.0-rc.20] - 2024-04-14
+
+- fix(plugins): Export plugin entrypoints on jsr.io
+- fix(core): Fix issue with PATH variable in process runner.
+
 ## [1.0.0-rc.19] - 2024-04-14
 
 - fix(core): Fix issue with PATH variable in process runner.
 - chore(core): Make code related to environment variables cross-runtime.
 - change(cli): Revert cli command `foreground` to `run`
 
-## [1.0.0-rc.19] - 2024-04-14
+## [1.0.0-rc.18] - 2024-04-14
 
 - fix(packaging): Fix regression bug in upgrader after moving to jsr.io